Move some settings to global
