+++ src/site/global/Global.vue
@@ -1,6 +1,9 @@
 <template>
 	<Tooltip />
 	<FloatContext />
+	<Transition name="settings-menu" appear>
+		<SettingsMenu v-if="settingsMenuCtx.open" />
+	</Transition>
 </template>
 
 <script setup lang="ts">
@@ -9,24 +12,35 @@ import { useConfig, useSettings } from "@/composable/useSettings";
 import useUpdater from "@/composable/useUpdater";
 import { useWorker } from "@/composable/useWorker";
 import FloatContext from "./FloatContext.vue";
+import { dataSettings, globalSettings } from "./GlobalSettings";
 import Tooltip from "./Tooltip.vue";
+import { useSettingsMenu } from "@/app/settings/Settings";
+import SettingsMenu from "@/app/settings/SettingsMenu.vue";
+
+const settingsMenuCtx = useSettingsMenu();
 
 const { register } = useSettings();
-register([
-	{
-		key: "app.version",
-		type: "NONE",
-		label: "App Version",
-		defaultValue: void 0 as never,
+
+register(dataSettings);
+register(globalSettings);
+
+const html = document.documentElement.classList;
+const body = document.body.classList;
+const useTransparency = useConfig("ui.transparent_backgrounds");
+
+watch(
+	useTransparency,
+	() => {
+		if (useTransparency.value) {
+			html.add("seventv-transparent");
+			body.add("seventv-transparent");
+		} else {
+			html.remove("seventv-transparent");
+			body.remove("seventv-transparent");
+		}
 	},
-	{
-		key: "app.7tv.token",
-		type: "NONE",
-		label: "7TV Bearer Token",
-		expires: 0,
-		defaultValue: "",
-	},
-]);
+	{ immediate: true },
+);
 
 const updater = useUpdater();
 const version = useConfig("app.version");
